Alice: Some Cleanup

+++ alice/plugins/forward.py
@@ -18,8 +18,7 @@ async def set_topics(c,m):
 	sql.add(channel_id,chat_id,thread_id)
 	await m.reply_text("Channel Connected")
 
-@Alice.on_message(filters.channel, group=1)
-async def forward(c,m):
+async def forward_m(c,m):
 	button = None
 	channel_id = m.chat.id
 	check = sql.check_channel(channel_id)
@@ -27,6 +26,9 @@ async def forward(c,m):
 		chat_id = check.chat_id
 		thread_id = check.thread_id
 		caption = ""
+		button = None
+		text = None
+		caption = None
 		if m.text:
 			with open('temp2.txt', 'w') as f:
 				f.write(m.text.html)
@@ -40,14 +42,15 @@ async def forward(c,m):
 				button = m.reply_markup
 			else:
 				button = None
-			os.remove(os.path.join(os.getcwd(), 'temp2.txt'))
+			if os.path.isfile(os.path.join(os.getcwd(), 'temp2.txt')):
+				os.remove(os.path.join(os.getcwd(), 'temp2.txt'))
 			return await c.send_message(chat_id, text, reply_to_message_id=thread_id, reply_markup=button, parse_mode=enums.ParseMode.HTML)
 		if m.caption:
 			with open('temp2.txt', 'w') as f:
 				f.write(m.caption.html)
 			file = open('temp2.txt', 'r')
 			caption = file.read()
-		caption, button = parse_button(caption)
+			caption, button = parse_button(caption)
 		if button:
 			button = build_keyboard(button)
 			button = InlineKeyboardMarkup(button)
@@ -68,4 +71,5 @@ async def forward(c,m):
 			await c.send_audio(chat_id=chat_id, audio=file, caption=caption, reply_to_message_id=thread_id, reply_markup=button, parse_mode=enums.ParseMode.HTML)
 		elif m.sticker:
 			await c.send_sticker(chat_id=chat_id, sticker=m.sticker.file_id, reply_to_message_id=thread_id)
-		os.remove(file)
+		if os.path.isfile(os.path.join(os.getcwd(), 'temp2.txt')):
+			os.remove(os.path.join(os.getcwd(), 'temp2.txt'))
