+++ alice/plugins/parser.py
@@ -1,15 +1,20 @@
 import os
 
 from alice.alice import Alice
+from alice.db import setting as sql
+from alice.plugins.forward import forward_m
 from alice.utils.string import build_keyboard, parse_button
 from pyrogram import enums, filters
 from pyrogram.types import InlineKeyboardMarkup
 
-@Alice.on_message(filters.channel, group=100)
+@Alice.on_message(filters.channel, group=1)
 async def channel_watcher(c,m):
 	chat_id = m.chat.id
-	if m.reply_markup:
-		return
+	forward = False
+	mess = m
+	check = sql.check_channel(chat_id)
+	if check:
+		forward = True
 	if m.text:
 		with open('temp.txt', 'w') as f:
 			f.write(m.text.html)
@@ -17,36 +22,38 @@ async def channel_watcher(c,m):
 		msg = file.read()
 		text, button = parse_button(msg)
 		button = build_keyboard(button)
-		if not button:
-			return os.remove(os.path.join(os.getcwd(), 'temp.txt'))
-		button = InlineKeyboardMarkup(button)
-		if m.forward_from:
-			await m.delete()
-			await c.send_message(chat_id=chat_id, text=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
-			return os.remove(os.path.join(os.getcwd(), 'temp.txt'))
-		await m.edit(text=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
-		os.remove(os.path.join(os.getcwd(), 'temp.txt'))
-
-	elif m.caption: # m.photo or m.video or m.animation:
+		if button:
+			button = InlineKeyboardMarkup(button)
+			if m.forward_from:
+				await m.delete()
+				mess = await c.send_message(chat_id=chat_id, text=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
+			else:
+				await m.edit(text=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
+	if m.caption: # m.photo or m.video or m.animation:
 		with open('temp.txt', 'w') as f:
 			f.write(m.caption.html)
 		file = open('temp.txt', 'r')
 		msg = file.read()
 		text, button = parse_button(msg)
 		button = build_keyboard(button)
-		if not button:
-			return os.remove(os.path.join(os.getcwd(), 'temp.txt'))
-		button = InlineKeyboardMarkup(button)
-		if m.forward_from:
-			if m.photo:
-				await c.send_photo(chat_id=chat_id, photo=m.photo.file_id, caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
-			elif m.video:
-				await c.send_video(chat_id=chat_id, video=m.video.file_id, caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
-			elif m.animation:
-				await c.send_animation(chat_id=chat_id, animation=m.animation.file_id, caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
-			await m.delete()
-			return os.remove(os.path.join(os.getcwd(), 'temp.txt'))
-		await m.edit_caption(caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
+		if button:
+			button = InlineKeyboardMarkup(button)
+			if m.forward_from:
+				if m.photo:
+					mess = await c.send_photo(chat_id=chat_id, photo=m.photo.file_id, caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
+				elif m.video:
+					mess = await c.send_video(chat_id=chat_id, video=m.video.file_id, caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
+				elif m.animation:
+					mess = await c.send_animation(chat_id=chat_id, animation=m.animation.file_id, caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
+				await m.delete()
+			else:
+				await m.edit_caption(caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
+	if forward:
+		try:
+			await mess.forward(chat_id=check.chat_id,message_thread_id=check.thread_id)
+		except Exception as e:
+			await forward_m(c,mess)
+	if os.path.isfile(os.path.join(os.getcwd(), 'temp.txt')):
 		os.remove(os.path.join(os.getcwd(), 'temp.txt'))
 
 @Alice.on_message(filters.private & filters.command("preview"))
@@ -85,4 +92,5 @@ async def preview(c,m):
 			await m.reply_video(video=m.video.file_id, caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
 		elif m.animation:
 			await m.reply_animation(animation=m.animation.file_id, caption=text, parse_mode=enums.ParseMode.HTML, reply_markup=button)
-		os.remove(os.path.join(os.getcwd(), 'temp.txt'))
+		if os.path.isfile(os.path.join(os.getcwd(), 'temp.txt')):
+			os.remove(os.path.join(os.getcwd(), 'temp.txt'))
